refactor: 댓글, 리뷰 리펙토링 진행  (#61)
* feat: 컨트롤러 swagger 추가 및 page 관련 제약 조건 추가

* feat: repository 및 request dto 제약조건 수정

* feat: metadata 관련 float -> double로 변경

* feat: review service 쪽 로직 수정

* refactor: 테스트 코드 수정

* feat: 레포지토리 인덱스 설명 추가

+++ src/main/java/com/threestar/trainus/domain/metadata/entity/ProfileMetadata.java
@@ -39,14 +39,14 @@ public class ProfileMetadata {
 
 	@Setter
 	@Column(nullable = false)
-	private Float rating;
+	private Double rating;
 
 	public void increaseReviewCount() {
 		this.reviewCount++;
 	}
 
-	public Float updateRating(Float newRating) {
-		Float totalRatingBefore = this.rating * (reviewCount - 1);
+	public double updateRating(double newRating) {
+		double totalRatingBefore = this.rating * (reviewCount - 1);
 		totalRatingBefore += newRating;
 		return totalRatingBefore / reviewCount;
 	}
